# design/IMS_RESTFull_API_Spec.yaml
openapi: 3.1.0
info:
  title: Inventory Management System API
  description: >
    REST API for the Inventory Management System (IMS). Designed for secure,
    auditable inventory and supplier management with role-based access control (RBAC),
    atomic stock transactions and barcode/QR workflows. This API
    prioritizes data integrity and operational simplicity for the v0.1 release.
    See externalDocs for full architecture, migrations and integration notes.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@inventory-system.com
  license:
    name: Proprietary
    url: https://company.com/license
servers:
  - url: https://api.inventory-system.com/v1
    description: Production server
  - url: https://staging-api.inventory-system.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoints
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "One or more validation errors occurred"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Email is required"
              code:
                type: string
                example: "required"
        trace_id:
          type: string
          example: "req_123456789"

    # Authentication Schemas
    User:
      type: object
      required:
        - id
        - email
        - full_name
        - role
        - is_active
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "admin@company.com"
        full_name:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [admin, inventory_manager, viewer]
          example: "admin"
        is_active:
          type: boolean
          example: true
        last_login_at:
          type: [ string, "null" ]
          format: date-time
          example: "2023-01-15T09:45:00Z"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-15T09:45:00Z"
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "admin@company.com"
        password:
          type: string
          format: password
          example: "securePassword123"
    LoginResponse:
      type: object
      required:
        - user
        - access_token
        - expires_in
      properties:
        user:
          $ref: '#/components/schemas/User'
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_in:
          type: integer
          example: 3600
        token_type:
          type: string
          example: "Bearer"
    PasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "user@company.com"
    PasswordResetConfirm:
      type: object
      required:
        - token
        - new_password
      properties:
        token:
          type: string
          example: "reset_token_123456"
        new_password:
          type: string
          format: password
          minLength: 8
          example: "newSecurePassword123"

    # Category Schemas
...
  responses:
    UnauthorizedError:
      description: Authentication credentials were missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "Authentication credentials missing or invalid"
            trace_id: "req_123456789"
...
    ValidationError:
      description: The request data failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "VALIDATION_ERROR"
            message: "One or more validation errors occurred"
            details:
              - field: "email"
                message: "Email is required"
                code: "required"
              - field: "password"
                message: "Password must be at least 8 characters"
                code: "min_length"
            trace_id: "req_123456789"
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Authenticate user and obtain access token
      description: Authenticate user credentials and return JWT token for API access (UC-01)
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "INVALID_CREDENTIALS"
                message: "Invalid email or password"
                trace_id: "req_123456789"
  /auth/logout:
    post:
      summary: Logout user and invalidate session
      description: Invalidate the current user session and access token
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully logged out"
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /auth/password-reset:
    post:
      summary: Request password reset
      description: Send password reset instructions to user's email
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        200:
          description: Password reset instructions sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset instructions sent to your email"
        400:
          $ref: '#/components/responses/ValidationError'
  /auth/password-reset/confirm:
    post:
      summary: Confirm password reset
      description: Reset user password using valid reset token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
      responses:
        200:
          description: Password successfully reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password successfully reset"
        400:
          $ref: '#/components/responses/ValidationError'
        410:
          description: Reset token expired or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "INVALID_RESET_TOKEN"
                message: "Reset token is invalid or has expired"
                trace_id: "req_123456789"

  # =============================================
  # USER MANAGEMENT ENDPOINTS (FR-AUTH-003 to FR-AUTH-006)
  # =============================================
...